updated installation via application.yaml
